Release: v1.1
I've changed the way I was handling the strlcpy function, because I
thought including the same bloated header I use on showip.c was stupid
when strlcpy only needs two. It also improves portability with the
addition of strlcpy.h.

Other notable change is the addition of an interface to handle input
from other programs, so you can pipe it for more complex filtering and
processing.

Finally, I reformated the output to something more compact, making it
easier to parse, with standard tools.