Build system update.
Full Make. I think this is the best option, for portability basically,
though it also simplifies the installation and possible binary
distribution. I removed little macro from the source too, but the version
stays the same. I'd like to try it on some BSDs next time, I think it's
going to work out of the box.